fix(track): segfault when decoding large dims
GK uses int32s for interval positions, but when multiplied with a dim component for storing/reading tracks, this can overflow (ie, reading/writing to negative positions).

There are several existing size_t promotions in the encode logic, so the simplest method is to do an onsite promote (rather than rework the logic to handel unsigned comparisons w/ underflow).