fix(appris/mane): allow data source to determine availability
Previously, GK hardcoded the filenames to encode what upstream version was used to build APPRIS/MANE. The issue with this is that older versions of GK could not be updated to use new APPRIS/MANE (ie, a new package with updated locator code needed to be released).

This change brings APPRIS/MANE in line with annotations: if they exist on the data source, then it's available to use in GK.

BREAKING CHANGE: this change also fixes gencode.v41 to use the correct source (e107 now instead of e103).